datamesh-ai-core/requirements.txt
# DATAMESH.AI Core Dependencies
# ==============================

# Core dependencies
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0.0,<7.0.0
jsonschema>=4.17.0,<5.0.0

# Tracing (OpenTelemetry)
opentelemetry-api>=1.20.0,<2.0.0
opentelemetry-sdk>=1.20.0,<2.0.0

# Optional: HTTP transport
# httpx>=0.24.0,<1.0.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.0.0
# ruff>=0.0.280
